Normalized spacing for the page so that extra spacing code was not in the card and left up to the page. Fixed some errors with the post filter and separated it into its own FilterBanner component.
